requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.11.17
cffi==1.16.0
circuitbreaker==1.4.0
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-import-export==3.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
flake8==7.0.0
MarkupPy==1.14
mccabe==0.7.0
oci==2.119.1
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.4
xlrd==2.0.1
xlwt==1.3.0
pyDes==2.0.1